// packages/core/src/router/Types.ts
import type { RouteMeta, RouteRecordRaw } from "vue-router";
import { defineComponent } from "vue";
import { Recordable } from "@mfish/types";

export type Component<T = any> =
  | ReturnType<typeof defineComponent>
  | (() => Promise<typeof import("*.vue")>)
  | (() => Promise<T>);

export interface AppRouteRecordRaw extends Omit<RouteRecordRaw, "children"> {
  name: string;
  meta: RouteMeta;
  redirect?: string;
  component?: Component | string;
  components?: Component;
  children?: AppRouteRecordRaw[];
  props?: Recordable;
  fullPath?: string;
  hideChildrenInMenu?: boolean;
}

export interface MenuTag {
  type?: "primary" | "error" | "warn" | "success";
  content?: string;
  dot?: boolean;
}

export interface Menu {
  name: string;
  icon?: string;
  path: string;
  // path contains param, auto assignment.
  paramPath?: string;
  disabled?: boolean;
  children?: Menu[];
  menuSort?: number;
  roles?: string[];
  meta?: Partial<RouteMeta>;
  tag?: MenuTag;
  hideMenu?: boolean;
  isExternal?: boolean;
  // 隐藏子菜单，当只有一级菜单时为true
  hideChildrenInMenu?: boolean;
}